Better handling of errors during feed refreshing
The Rake tasks output more useful information and try to store the
offending XML for inspection. The core refresher now raises a wrapped
error so that all types of errors can be caught by the rake task and the
worker, allowing the process to keep running.

Problematic feeds still need to be marked to avoid worker churn.

Also a note about worker concurrency; it doesn't work yet.